requirements.txt
asgiref==3.2.10
awsebcli==3.18.1
Babel==2.8.0
botocore==1.15.49
cement==2.8.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==2.2.14
django-appconf==1.0.4
django-autoslug==1.9.8
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-image-optimizer==1.0.1
django-imagekit==2.0.4
django-js-asset==1.2.2
django-optimized-image==0.3.0
django-phonenumber-field==4.0.0
django-user-agents==0.4.0
djangorestframework==3.11.1
docutils==0.15.2
future==0.16.0
idna==2.7
jmespath==0.10.0
MouseInfo==0.1.3
numpy==1.19.1
pathspec==0.5.9
phonenumbers==8.12.6
Pillow==7.1.2
psycopg2==2.8.5
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.0
python-memcached==1.59
python-resize-image==1.1.19
PyTweening==1.0.3
pytz==2020.1
pywin32==227
PyYAML==5.3.1
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
termcolor==1.1.0
tinify==1.5.1
ua-parser==0.10.0
urllib3==1.24.3
user-agents==2.1
wcwidth==0.1.9
